Code clean up and tool updates
* Imagine: Clean up util/utility.h macros
* Imagine: Use __builtin_assume() if supported
* Imagine: Make PixmapDesc a struct
* Imagine: Allow returning a value in doIfUsed()
* Imagine: Fix buffer size values in PipeMessagePort::sendWithExtraData()
* Imagine: Update gradle to 7.5.1 and bundletool to 1.12.1
* EmuFramework: Add a link to the online documentation in the options menu
* EmuFramework: Use a new enum value for autosave launch mode option
* Update Github CI workflow and remove stale PR/issue job

+++ imagine/src/gui/ViewStack.cc
@@ -359,7 +359,7 @@ View &ViewStack::top() const
 
 View &ViewStack::viewAtIdx(int idx) const
 {
-	assumeExpr((unsigned)idx < view.size());
+	assumeExpr(size_t(idx) < view.size());
 	return *view[idx].v;
 }
